refactor(watch): carve rebuild.rs out of mod.rs
src/cli/watch/rebuild.rs (~470 lines):
- hnsw_rebuild_threshold (env-var read)
- PendingRebuild + RebuildResult + RebuildOutcome + MAX_PENDING_REBUILD_DELTA
- EmbedderBackoff + impl
- try_init_embedder, resolve_index_aware_model_for_watch
- spawn_hnsw_rebuild, drain_pending_rebuild, clear_hnsw_dirty_with_retry

PendingRebuild fields lifted to pub(super) where cmd_watch and tests touch
them (delta, delta_saturated, handle, rx, started_at). EmbedderBackoff
fields/methods lifted similarly so the existing test bench keeps reaching
them via 'use super::*;' in mod tests.
